[OBEY-CAMPAIGN-8deed8b4] fix: preserve divergent scaffold files during migration

+++ internal/intent/migration_filesystem.go
@@ -147,7 +147,13 @@ func collectIntentTreeMoves(srcDir, dstDir string, moves *[]PlannedPathMove) err
 
 		if _, err := os.Stat(dstPath); err == nil {
 			if isIntentScaffoldBasename(entry.Name()) {
-				continue
+				matches, matchErr := scaffoldFilesMatch(srcPath, dstPath)
+				if matchErr != nil {
+					return matchErr
+				}
+				if matches {
+					continue
+				}
 			}
 			return camperrors.Wrapf(ErrIntentMigrationConflict, "destination already exists for %s", dstPath)
 		} else if !os.IsNotExist(err) {
@@ -264,10 +270,16 @@ func moveIntentTree(srcDir, dstDir string) error {
 
 		if _, err := os.Stat(dstPath); err == nil {
 			if isIntentScaffoldBasename(entry.Name()) {
-				if err := os.Remove(srcPath); err != nil {
-					return camperrors.Wrapf(err, "removing %s", srcPath)
+				matches, matchErr := scaffoldFilesMatch(srcPath, dstPath)
+				if matchErr != nil {
+					return matchErr
 				}
-				continue
+				if matches {
+					if err := os.Remove(srcPath); err != nil {
+						return camperrors.Wrapf(err, "removing %s", srcPath)
+					}
+					continue
+				}
 			}
 			return camperrors.Wrapf(ErrIntentMigrationConflict, "destination already exists for %s", dstPath)
 		} else if !os.IsNotExist(err) {
@@ -284,4 +296,18 @@ func moveIntentTree(srcDir, dstDir string) error {
 	}
 
 	return nil
+}
+
+func scaffoldFilesMatch(srcPath, dstPath string) (bool, error) {
+	srcData, err := os.ReadFile(srcPath)
+	if err != nil {
+		return false, camperrors.Wrapf(err, "reading %s", srcPath)
+	}
+
+	dstData, err := os.ReadFile(dstPath)
+	if err != nil {
+		return false, camperrors.Wrapf(err, "reading %s", dstPath)
+	}
+
+	return bytes.Equal(srcData, dstData), nil
 }
